requirements.txt
aiohttp
django
azure-storage-blob
azure-storage-file
azure-storage-queue
azure-storage-common
